requirements.txt
streamlit
newspaper3k
langchain
langchain-community
googletrans==4.0.0-rc1
gTTS
ollama
